Refactor: Extract holiday-service.js from calendar.js
