+++ src/main/javascript/components/calendar/index.js
@@ -1,5 +1,5 @@
 import { Calendar } from "./calendar";
-import { HolidayService } from "./calendar";
+import { HolidayService } from "./holiday-service";
 
 /**
  * @export
